+++ src/components/Favorites.js
@@ -1,8 +1,7 @@
 import React, { Component } from 'react';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
-import ItemList from './ItemList';
-import Search from './Search';
+import { ItemList, Search } from './';
 import { getFavoriteItems } from '../actions';
 
 class Favorites extends Component {
