+++ src/components/index.js
@@ -1,0 +1,15 @@
+import Details from './Details';
+import Favorites from './Favorites';
+import Home from './Home';
+import Item from './Item';
+import ItemList from './ItemList';
+import Search from './Search';
+
+export {
+  Details,
+  Favorites,
+  Home,
+  Item,
+  ItemList,
+  Search
+}
