+++ src/components/styled/index.js
@@ -1,0 +1,11 @@
+import FavButton from './FavButton';
+import SearchButton from './SearchButton';
+import SearchDiv from './SearchDiv';
+import SearchText from './SearchText';
+
+export {
+  FavButton,
+  SearchButton,
+  SearchDiv,
+  SearchText
+}
